jd_ingestion_system/app/core/config.py
from functools import lru_cache
from pathlib import Path

from dotenv import load_dotenv
from pydantic_settings import BaseSettings, SettingsConfigDict


load_dotenv()


class Settings(BaseSettings):
    app_name: str = "JD Ingestion System"
    use_llm_segmentation: bool = False
    data_dir: Path = Path("data")
    raw_dir: Path = Path("data/raw")
    processed_dir: Path = Path("data/processed")
    metadata_path: Path = Path("data/metadata.json")
    faiss_index_path: Path = Path("data/faiss.index")
    vector_metadata_path: Path = Path("data/vector_metadata.json")
    max_file_size_bytes: int = 10 * 1024 * 1024
    embedding_model_name: str = "all-MiniLM-L6-v2"

    model_config = SettingsConfigDict(env_file=".env", env_file_encoding="utf-8")


@lru_cache(maxsize=1)
def get_settings() -> Settings:
    settings = Settings()
    settings.raw_dir.mkdir(parents=True, exist_ok=True)
    settings.processed_dir.mkdir(parents=True, exist_ok=True)
    settings.data_dir.mkdir(parents=True, exist_ok=True)
    return settings
