jd_ingestion_system/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic-settings==2.5.2
python-multipart==0.0.9
python-docx==1.1.2
PyMuPDF==1.24.10
pytesseract==0.3.13
Pillow==10.4.0
trafilatura==1.12.2
beautifulsoup4==4.12.3
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
numpy==1.26.4
requests==2.32.3
